requirements_current.txt
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.57.0
fsspec==2024.6.1
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.0.3
pillow==10.2.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1+cpu
torchvision==0.19.1+cpu
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
zipp==3.20.2
